Refactor event handling code
