+++ src/routes/[eventUrl]/_components/calendar/actions/calendarInteraction.js
@@ -1,52 +1,7 @@
 import dayjs from 'dayjs';
 
+import { getTarget } from 'src/utils/eventHandler';
 import LongTouchAndDrag from 'src/utils/LongTouchAndDrag';
-import { MS_PER_HOUR } from 'src/utils/constants';
-
-/**
- * Get the target of the event, regardless of whether it was a mouse or touch
- * event.
- * @param {MouseEvent | TouchEvent} event The mouse or touch event.
- * @returns {Element} The target of the event.
- */
-function getTarget(event) {
-  if (event instanceof TouchEvent) {
-    return getTouchTarget(event);
-  }
-  return event.target;
-}
-
-function getTouchTarget(event) {
-  const { clientX, clientY } = event.changedTouches[0];
-  return document.elementFromPoint(clientX, clientY);
-}
-
-function getOffset(event) {
-  if (event instanceof TouchEvent) {
-    return getTouchOffset(event);
-  }
-  return getMouseOffset(event);
-}
-
-function getTouchOffset(event) {
-  const { target } = event;
-  const rect = target.getBoundingClientRect();
-  const touch = event.targetTouches[0];
-  const offsetX = touch.clientX - rect.left;
-  const offsetY = touch.clientY - rect.top;
-  return { offsetX, offsetY };
-}
-
-function getMouseOffset(event) {
-  // Cross-browser calculation of offsetY by Jack Moore, 2012.
-  // https://www.jacklmoore.com/notes/mouse-position/
-  // event.offsetX and event.offsetY breaks on Safari when zooming the canvas in.
-  const { target } = event;
-  const rect = target.getBoundingClientRect();
-  const offsetX = event.clientX - rect.left;
-  const offsetY = event.clientY - rect.top;
-  return { offsetX, offsetY };
-}
 
 function isQuarterHourTarget(target) {
   return target.dataset.quarterHourTarget != null;
